pyproject.toml
[project]
name = "osu-mapper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.7.0",
    "audioop-lts>=0.2.1",
    "einops>=0.8.1",
    "flask>=3.1.1",
    "gym>=0.26.2",
    "hydra-core>=1.3.2",
    "lightning>=2.5.1.post0",
    "matplotlib>=3.10.3",
    "nnaudio>=0.3.3",
    "pandas>=2.2.3",
    "pyarrow>=20.0.0",
    "pydub>=0.25.1",
    "pywebview[qt]>=5.4",
    "pyyaml>=6.0.2",
    "slider",
    "tensorboard>=2.19.0",
    "torch>=2.7.0",
    "torch-tb-profiler>=0.4.3",
    "torchaudio>=2.7.0",
    "transformers>=4.52.4",
    "wandb>=0.19.11",
]

[tool.uv.sources]
slider = { git = "https://github.com/OliBomby/slider.git", rev = "gedagedigedagedaoh" }
